SL-54: minor fixes
