+++ src/helpers/common.js
@@ -88,6 +88,47 @@ const randomUniqueId = () => {
   return timeStamp + randomString + Math.floor(Math.random() * 999);
 };
 
+const processImage = (file, callback) => {
+  const reader = new FileReader();
+  reader.onload = () => {
+    const img = new Image();
+    img.onload = () => {
+      const MAX_WIDTH = 800;
+      const MAX_HEIGHT = 600;
+
+      const canvas = document.createElement('canvas');
+      const ctx = canvas.getContext('2d');
+
+      let width = img.width;
+      let height = img.height;
+
+      if (width > height) {
+        if (width > MAX_WIDTH) {
+          height *= MAX_WIDTH / width;
+          width = MAX_WIDTH;
+        }
+      } else {
+        if (height > MAX_HEIGHT) {
+          width *= MAX_HEIGHT / height;
+          height = MAX_HEIGHT;
+        }
+      }
+
+      canvas.width = width;
+      canvas.height = height;
+      ctx.drawImage(img, 0, 0, width, height);
+
+      const compressedImageUrl = canvas.toDataURL('image/jpeg', 0.7);
+
+      callback(compressedImageUrl);
+    };
+
+    img.src = reader.result;
+  };
+
+  reader.readAsDataURL(file);
+};
+
 export {
   goHome,
   scrollToTop,
@@ -99,4 +140,5 @@ export {
   sortByState,
   getCurrentDate,
   randomUniqueId,
+  processImage,
 };
